fix all remarks
