--- src/containers/guest-home-page/how-it-works/studentCardBoxArray.js
@@ -1,27 +1,0 @@
-import howItWorksStudentFirst from '~/assets/img/guest-home-page/howItWorksStudentFirst.svg'
-import howItWorksStudentSecond from '~/assets/img/guest-home-page/howItWorksStudentSecond.svg'
-import howItWorksStudentThird from '~/assets/img/guest-home-page/howItWorksStudentThird.svg'
-import howItWorksStudentFourth from '~/assets/img/guest-home-page/howItWorksStudentFourth.svg'
-
-export const studentCardBoxArray = [
-  {
-    image: howItWorksStudentFirst,
-    title: 'guestHomePage.howItWorks.student.signUp.title',
-    description: 'guestHomePage.howItWorks.student.signUp.description'
-  },
-  {
-    image: howItWorksStudentSecond,
-    title: 'guestHomePage.howItWorks.student.selectAMentor.title',
-    description: 'guestHomePage.howItWorks.student.selectAMentor.description'
-  },
-  {
-    image: howItWorksStudentThird,
-    title: 'guestHomePage.howItWorks.student.sendRequest.title',
-    description: 'guestHomePage.howItWorks.student.sendRequest.description'
-  },
-  {
-    image: howItWorksStudentFourth,
-    title: 'guestHomePage.howItWorks.student.startLearning.title',
-    description: 'guestHomePage.howItWorks.student.startLearning.description'
-  }
-]
